% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corpBT17_20}
\alias{corpBT17_20}
\title{Original Data used to calculate weights}
\format{
A corpus .
}
\usage{
corpBT17_20
}
\description{
Sanmple textdata from the 17. till the ongoing 20. period (2023) of the German Parliament.
}
\references{
Klotz, T. (2023). How can populism be identified in political texts taking context into account?: Development of a new populism dictionary
\href{https://github.com/TomSOWI/BA-Arbeit}

Open Discourse (2023)
\href{https://github.com/open-discourse/open-discourse}
}
\keyword{datasets}
