% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_tokens_updated.R
\name{get_pop_tokens_updated}
\alias{get_pop_tokens_updated}
\title{Title}
\usage{
get_pop_tokens_updated(
  text,
  create_compounds = FALSE,
  compounds_at_level = "sentences",
  compounds_dict = NULL,
  compounds_dict_glob = FALSE
)
}
\arguments{
\item{compounds_dict_glob}{}
}
\description{
Title
}
