% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klotz_clara}
\alias{klotz_clara}
\title{Best performed dictionary by klotz using clara-clustering.}
\format{
A data frame.
}
\usage{
klotz_clara
}
\description{
A dataframe including gruendls populist terms and klotzs patterns weights created with the clara-algorithm
}
\references{
Klotz, T. (2023). How can populism be identified in political texts taking context into account?: Development of a new populism dictionary
\href{https://github.com/TomSOWI/BA-Arbeit}

Gründl, J. (2020). Populist ideas on social media: A dictionary-based
measurement of populist communication. \emph{New Media & Society}. Advance
online publication.
\href{https://doi.org/10.1177/1461444820976970}{https://doi.org/10.1177/1461444820976970}
}
\keyword{datasets}
