% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klotz_clara_DACH}
\alias{klotz_clara_DACH}
\title{klotz_clara including weighted patterns for DACH region.}
\format{
A data frame.
}
\usage{
klotz_clara_DACH
}
\description{
Since the klotz_clara dictionary was trained on data from germany patterns addressing austria and swiss are not included in the original dictionary.
This version includes patterns for the DACH region with its corresponding weights addressing germans
(zu Lasten der Deutschen -> zu Lasten der Schweizer)
}
\references{
Klotz, T. (2023). How can populism be identified in political texts taking context into account?: Development of a new populism dictionary
\href{https://github.com/TomSOWI/BA-Arbeit}

Gründl, J. (2020). Populist ideas on social media: A dictionary-based
measurement of populist communication. \emph{New Media & Society}. Advance
online publication.
\href{https://doi.org/10.1177/1461444820976970}{https://doi.org/10.1177/1461444820976970}
}
\keyword{datasets}
