% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klotz_clara_N}
\alias{klotz_clara_N}
\title{klotz_clara including the number of cases used for weight calculation}
\format{
A data frame.
}
\usage{
klotz_clara_N
}
\description{
As mentioned in the paper Klotz 2023 the low amount of cases for some patterns could be an issue.
This dataframe allows to filter those cases using the colomn match_count to further explore different versions of klotz_clara.
}
\references{
Klotz, T. (2023). How can populism be identified in political texts taking context into account?: Development of a new populism dictionary
\href{https://github.com/TomSOWI/BA-Arbeit}

Gründl, J. (2020). Populist ideas on social media: A dictionary-based
measurement of populist communication. \emph{New Media & Society}. Advance
online publication.
\href{https://doi.org/10.1177/1461444820976970}{https://doi.org/10.1177/1461444820976970}
}
\keyword{datasets}
