% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_patterns_compound.R
\name{make_patterns_compound}
\alias{make_patterns_compound}
\title{Optimizes regex-patterns for deployment on compound textdata}
\usage{
make_patterns_compound(pattern)
}
\arguments{
\item{pattern}{A character vector}
}
\value{
A character vector
}
\description{
Optimizes regex-patterns for deployment on compound textdata
}
\examples{
make_patterns_compound("zum schaden de(s|r)")
}
