% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-weightdict.R
\name{run_weightdict}
\alias{run_weightdict}
\title{Deploy a weighted dictionary as a dataframe to text}
\usage{
run_weightdict(
  text,
  dict,
  dict_name = "weighted_dict",
  prepare_corp = FALSE,
  add_no_weight = FALSE
)
}
\arguments{
\item{text}{Data should be provided as an object that can be used for quntedas kwic function.
Note: If you want to apply a dictionary containing wildcard patterns then you should prepare your textdata using corpus_to_compound_tokens().
Otherwise you can set prepare_corp = TRUE but this is not recommended since it will take a while.}

\item{dict}{Provide a dataframe containing a column "pattern" and "weight"}

\item{dict_name}{Set a custom name}

\item{prepare_corp}{Set to TRUE if you apply a dictionary containing wildcards and you have not prepared your textdata}

\item{add_no_weight}{Should results without considering weights be returned in addition}
}
\value{
A dataframe objecct including the populism score per document
}
\description{
Deploy a weighted dictionary as a dataframe to text
}
\examples{
run_weightdict(
text = toksBT17_20_clean,
dict = klotz_clara,
dict_name = "klotz"
)
}
