% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-weightdict.R
\name{run_weightdictR}
\alias{run_weightdictR}
\title{Title}
\usage{
run_weightdictR(
  data,
  dict_compound,
  dict_name = "original",
  dict_weight,
  dict_weight_name = "weighted",
  prepare_corp = FALSE,
  include_main_dict = TRUE,
  include_totals = TRUE
)
}
\arguments{
\item{include_totals}{}
}
\description{
Title
}
