% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toksBT17_20_clean}
\alias{toksBT17_20_clean}
\title{Data used to calculate weights}
\format{
A token .
}
\usage{
toksBT17_20_clean
}
\description{
Textdata from the 17. till the ongoing 20. period (2023) of the German Parliament.
This data is already to run my weighted dictionary klotz_clara
}
\references{
Klotz, T. (2023). How can populism be identified in political texts taking context into account?: Development of a new populism dictionary
\href{https://github.com/TomSOWI/BA-Arbeit}

Open Discourse (2023)
\href{https://github.com/open-discourse/open-discourse}
}
\keyword{datasets}
