% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_to_compound_tokens.R
\name{corpus_to_compound_tokens}
\alias{corpus_to_compound_tokens}
\title{Title}
\usage{
corpus_to_compound_tokens(corp, pattern)
}
\arguments{
\item{corp}{A corpus object}

\item{pattern}{A list of patterns or a quanteda dictionary object}
}
\description{
Title
}
