% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pattern_stats.R
\name{get_pattern_stats}
\alias{get_pattern_stats}
\title{Get information on how many times a pattern occurred in a text}
\usage{
get_pattern_stats(text, pattern, regex = TRUE)
}
\arguments{
\item{text}{A corpus or tokens object suitable for quantedas kwic function}

\item{pattern}{A dictionary or a character vector}

\item{regex}{Do the patterns contain regex? TRUE | FALSE}
}
\value{
A dataframe with the columns pattern and match_count
}
\description{
Get information on how many times a pattern occurred in a text
}
\examples{
get_pattern_stats(text = toksBT17_20_clean, pattern = klotz_clara$pattern, regex = TRUE)
}
